+++ src/components/Table/Doctor/DoctorsList.tsx
@@ -2,6 +2,7 @@ import { Tab, TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react";
 import ApprovedDocList from "./ApprovedDocList";
 import PendingDocList from "./PendingDocList";
 import RegisteredDocList from "./RegisteredDocList";
+import RejectedDocList from "./RejectedDocList";
 
 const DoctorsList = () => {
   return (
@@ -10,6 +11,7 @@ const DoctorsList = () => {
         <Tab>Registered Doctors</Tab>
         <Tab>Pending Approval</Tab>
         <Tab>Approved</Tab>
+        <Tab>Rejected</Tab>
       </TabList>
       <TabPanels>
         <TabPanel>
@@ -20,6 +22,9 @@ const DoctorsList = () => {
         </TabPanel>
         <TabPanel>
           <ApprovedDocList />
+        </TabPanel>
+        <TabPanel>
+          <RejectedDocList />
         </TabPanel>
       </TabPanels>
     </Tabs>